global exception handling

+++ src/main/java/org/springMvc/service/EmpServiceImpl.java
@@ -2,6 +2,7 @@ package org.springMvc.service;
 
 import java.util.List;
 
+import org.springMvc.exception.EmpNotFound;
 import org.springMvc.model.Emp;
 import org.springMvc.repo.EmpRepo;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,6 +18,13 @@ public class EmpServiceImpl implements EmpService{
 	}
 	@Override
 	public List<Emp> display() {
+		
+		 List<Emp> list = empRepo.display();
+
+		    if(list.isEmpty()) {
+		        throw new EmpNotFound("No employees found in database");
+		    }
+
 		return empRepo.display();
 	}
 	@Override
@@ -26,10 +34,21 @@ public class EmpServiceImpl implements EmpService{
 	}
 	@Override
 	public void delete(int eid) {
+		
+		List<Emp>list=empRepo.search(eid);
+		if(list.isEmpty())
+		{
+			throw new EmpNotFound("emp with "+eid+" not found");
+		}
 		empRepo.delete(eid);
 	}
 	@Override
 	public void update(Emp e) {
+		List<Emp>list=empRepo.search(e.getEid());
+		if(list.isEmpty())
+		{
+			throw new EmpNotFound("emp with "+e.getEid()+" not found");
+		}
 		empRepo.update(e);
 	}
